+++ src/components/dashboard/index.ts
@@ -16,3 +16,4 @@ export {
 export { EmptyState, LoadingState, ScaffoldNote } from "./States";
 export { default as DataTable, type Column } from "./DataTable";
 export { default as DashboardShell } from "./DashboardShell";
+export { default as PipelineBoard } from "./PipelineBoard";